chore: Move swipe control gesture to separate component
- This way, if it's triggering accidentally too frequenty, we can scope it by wrapping the artwork & metadata portion of the "Now Playing" screen instead of doing it everwhere.
