+++ mobile/src/navigation/screens/now-playing/View.tsx
@@ -5,17 +5,9 @@ import { useNavigation } from "@react-navigation/native";
 import type { NativeStackNavigationProp } from "@react-navigation/native-stack";
 import { useTranslation } from "react-i18next";
 import { View } from "react-native";
-import {
-  Directions,
-  GestureDetector,
-  useCompetingGestures,
-  useFlingGesture,
-} from "react-native-gesture-handler";
-import { scheduleOnRN } from "react-native-worklets";
 
 import type { Track } from "~/data/track/types";
 import { usePlaybackStore } from "~/stores/Playback/store";
-import { PlaybackControls } from "~/stores/Playback/actions";
 import { usePreferenceStore } from "~/stores/Preference/store";
 import { presentTrackSheet } from "~/stores/Session/actions";
 import { toggleLyricVisibility } from "~/modules/lyric/core/actions";
@@ -42,32 +34,16 @@ import {
   RepeatButton,
   ShuffleButton,
 } from "~/modules/media/components/MediaControls";
+import { PlaybackControlGestureWrapper } from "./components/PlaybackControlGestureWrapper";
 import { FavoriteButton } from "~/modules/media/components/Track";
 
 export default function NowPlaying() {
   const track = usePlaybackStore((s) => s.activeTrack);
-
-  //#region Swipe Control Gestures
-  const enableSwipeGesture = usePreferenceStore((s) => s.nowPlayingGestures);
-
-  const swipeLeftGesture = useFlingGesture({
-    enabled: enableSwipeGesture,
-    direction: Directions.LEFT,
-    onActivate: () => scheduleOnRN(PlaybackControls.next),
-  });
-  const swipeRightGesture = useFlingGesture({
-    enabled: enableSwipeGesture,
-    direction: Directions.RIGHT,
-    onActivate: () => scheduleOnRN(PlaybackControls.prev),
-  });
-  const gestures = useCompetingGestures(swipeLeftGesture, swipeRightGesture);
-  //#endregion
-
   if (!track) return <Back />;
   return (
     <SeekbarContext>
       <SafeContainer additionalTopOffset={56} className="flex-1 gap-8">
-        <GestureDetector gesture={gestures}>
+        <PlaybackControlGestureWrapper>
           <ArtworkSlot artwork={track.artwork} trackId={track.id} />
           <View className="-mt-8 gap-6 px-4">
             <Metadata track={track} />
@@ -76,10 +52,10 @@ export default function NowPlaying() {
               uri={track.uri}
               trackLength={track.duration}
             />
-            <PlaybackMediaControls />
+            <PlaybackControls />
           </View>
           <BottomAppBar trackId={track.id} />
-        </GestureDetector>
+        </PlaybackControlGestureWrapper>
       </SafeContainer>
     </SeekbarContext>
   );
@@ -138,7 +114,7 @@ function Metadata({ track }: { track: Track }) {
 //#endregion
 
 //#region Playback Controls
-function PlaybackMediaControls() {
+function PlaybackControls() {
   return (
     <View className="mx-auto w-full max-w-96 flex-row items-center justify-between gap-2 rtl:flex-row-reverse">
       <ShuffleButton />
